3 часть

+++ src/main/java/ru/skypro/homework/dto/AdDto.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class AdDto {
-    private int author; //id автора объявления
+    private Long author; //id автора объявления
 
     private String image;//ссылка на картинку объявления
